Abstracting common parts for image transforming actions
